third_party/blink/web_tests/flag-specific/disable-layout-ng/fast/table/caption-margin-direction-expected.txt

This is a testharness.js-based test.
FAIL Test caption location 0 assert_equals: expected 20 but got 0
PASS Test caption location 1
Harness: the test ran to completion.
